Add files via upload
To test the receiver's signal make sure it is bound to your transmitter, + goes to arduino's 5V pin, and - goes to arduino's ground. Upload the code to your arduino and connect the signal pin to the receiving pin in the arduino (pin 2 in this code). Open your serial monitor and your signal will be displayed.